+++ src/model/game/projectile/Plasmaball.h
@@ -8,12 +8,12 @@ class Plasmaball
 private:
     bool player;
     uint_fast32_t damages;
-    int_fast32_t xVelocity;
-    int_fast32_t yVelocity;
+    long double xVelocity;
+    long double yVelocity;
     long double x;
     long double y;
 public:
-    Plasmaball(bool isPlayer, uint_fast32_t damages, int_fast32_t xVelocity, int_fast32_t yVelocity, long double x, long double y);
+    Plasmaball(bool isPlayer, uint_fast32_t damages, long double xVelocity, long double yVelocity, long double x, long double y);
     void move(long double lag);
     bool isPlayer() const;
     uint_fast32_t getDamages() const;
